git-svn-id: http://astropew.googlecode.com/svn/trunk@184 d889380e-1d37-11de-8b5c-13accb87f508

+++ src/client/AbstractCommand.java
@@ -2,13 +2,13 @@ package client;
 
 public abstract class AbstractCommand implements Command {
 
-	protected long time;
+	protected float time;
 
-	public AbstractCommand(long time) {
+	public AbstractCommand(float time) {
 		this.time = time;
 	}
 
-	public long getTime() {
+	public float getTime() {
 		return time;
 	}
