+++ src/client/Command.java
@@ -5,6 +5,6 @@ public interface Command extends Comparable<Command> {
 
 	public void perform(Game game, float interpolation);
 
-	public long getTime();
+	public float getTime();
 
 }
