+++ src/server/FireMissileCommand.java
@@ -6,7 +6,7 @@ public class FireMissileCommand extends AbstractCommand {
 		
 	SocketAddress sender;
 	
-	public FireMissileCommand(SocketAddress sender, long t) {
+	public FireMissileCommand(SocketAddress sender, float t) {
 		super(t);
 		this.sender = sender;
 	}
